rn: Move some login actions into better files
