+++ packages/edge-login-ui-rn/src/actions/LoginAction.js
@@ -1,13 +1,6 @@
 // @flow
 
-import { makeReactNativeDisklet } from 'disklet'
-import React from 'react'
-import { sprintf } from 'sprintf-js'
-
 import s from '../common/locales/strings.js'
-import { ButtonsModal } from '../components/modals/ButtonsModal.js'
-import { Airship } from '../components/services/AirshipInstance.js'
-import * as Constants from '../constants/index.js'
 import {
   enableTouchId,
   isTouchDisabled,
@@ -17,6 +10,8 @@ import {
 } from '../keychain.js'
 import type { Dispatch, GetState, Imports } from '../types/ReduxTypes.js'
 import { translateError } from '../util/ErrorMessageUtil.js'
+import { twofaReminder } from './LoginCompleteActions.js'
+import { setMostRecentUsers } from './PreviousUsersActions.js'
 
 /**
  * Make it Thunky
@@ -371,193 +366,4 @@ export function recoveryLoginComplete() {
     dispatch({ type: 'CLOSE_NOTIFICATION_MODAL' })
     callback(null, account, touchIdInformation)
   }
-}
-// validateUsername check
-
-const twofaReminder = async account => {
-  const { otpKey, dataStore } = account
-  const pluginList = await dataStore.listStoreIds()
-  const storeName = pluginList.includes(Constants.OTP_REMINDER_STORE_NAME)
-    ? Constants.OTP_REMINDER_STORE_NAME
-    : null
-  const itemList = storeName
-    ? await dataStore.listItemIds(Constants.OTP_REMINDER_STORE_NAME)
-    : null
-  const createdAtString =
-    itemList && itemList.includes(Constants.OTP_REMINDER_KEY_NAME_CREATED_AT)
-      ? await dataStore.getItem(
-          Constants.OTP_REMINDER_STORE_NAME,
-          Constants.OTP_REMINDER_KEY_NAME_CREATED_AT
-        )
-      : null
-  const createdAt = createdAtString ? parseInt(createdAtString) : null
-  const reminderCreatedAtDate = createdAt
-    ? createdAt + Constants.OTP_REMINDER_MILLISECONDS
-    : null
-  const lastOtpCheckString =
-    itemList &&
-    itemList.includes(Constants.OTP_REMINDER_KEY_NAME_LAST_OTP_CHECKED)
-      ? await dataStore.getItem(
-          Constants.OTP_REMINDER_STORE_NAME,
-          Constants.OTP_REMINDER_KEY_NAME_LAST_OTP_CHECKED
-        )
-      : null
-  const lastOtpCheck = lastOtpCheckString ? parseInt(lastOtpCheckString) : null
-  const reminderLastOtpCheckDate = lastOtpCheck
-    ? lastOtpCheck + Constants.OTP_REMINDER_MILLISECONDS
-    : null
-  const dontAsk =
-    itemList && itemList.includes(Constants.OTP_REMINDER_KEY_NAME_DONT_ASK)
-      ? await dataStore.getItem(
-          Constants.OTP_REMINDER_STORE_NAME,
-          Constants.OTP_REMINDER_KEY_NAME_DONT_ASK
-        )
-      : null
-
-  const enableOtp = async account => {
-    await account.enableOtp()
-    return Airship.show(bridge => (
-      <ButtonsModal
-        bridge={bridge}
-        title={s.strings.otp_authentication_header}
-        message={sprintf(s.strings.otp_authentication_message, account.otpKey)}
-        buttons={{ ok: { label: s.strings.ok } }}
-      />
-    ))
-  }
-
-  const createOtpCheckModal = async () => {
-    const result = await Airship.show(bridge => (
-      <ButtonsModal
-        bridge={bridge}
-        title={s.strings.otp_reset_modal_header}
-        message={s.strings.otp_reset_modal_message}
-        buttons={{
-          yes: { label: s.strings.enable },
-          no: { label: s.strings.skip_button, type: 'secondary' }
-        }}
-      />
-    ))
-    return result === 'yes'
-  }
-
-  const createOtpCheckModalDontAsk = async () => {
-    return Airship.show(bridge => (
-      <ButtonsModal
-        bridge={bridge}
-        header={s.strings.otp_reset_modal_header}
-        message={s.strings.otp_reset_modal_message}
-        buttons={{
-          enable: { label: s.strings.enable, type: 'primary' },
-          cancel: { label: s.strings.skip_button, type: 'secondary' },
-          dontAsk: {
-            label: s.strings.otp_reset_modal_dont_ask,
-            type: 'secondary'
-          }
-        }}
-      />
-    ))
-  }
-
-  if (otpKey) {
-    return true
-  }
-
-  if (dontAsk) {
-    return true
-  }
-
-  if (!storeName) {
-    const resolve = await createOtpCheckModal()
-    if (resolve) {
-      await enableOtp(account)
-      await account.dataStore.setItem(
-        Constants.OTP_REMINDER_STORE_NAME,
-        Constants.OTP_REMINDER_KEY_NAME_LAST_OTP_CHECKED,
-        Date.now().toString()
-      )
-      return true
-    }
-    await account.dataStore.setItem(
-      Constants.OTP_REMINDER_STORE_NAME,
-      Constants.OTP_REMINDER_KEY_NAME_LAST_OTP_CHECKED,
-      Date.now().toString()
-    )
-    return false
-  }
-
-  if (
-    lastOtpCheckString &&
-    lastOtpCheck &&
-    reminderLastOtpCheckDate &&
-    Date.now() > reminderLastOtpCheckDate
-  ) {
-    const resolve = await createOtpCheckModalDontAsk()
-    if (resolve === 'enable') {
-      await enableOtp(account)
-      return true
-    }
-    if (resolve === 'dontAsk') {
-      await account.dataStore.setItem(
-        Constants.OTP_REMINDER_STORE_NAME,
-        Constants.OTP_REMINDER_KEY_NAME_DONT_ASK,
-        'true'
-      )
-      return false
-    }
-    await account.dataStore.setItem(
-      Constants.OTP_REMINDER_STORE_NAME,
-      Constants.OTP_REMINDER_KEY_NAME_LAST_OTP_CHECKED,
-      Date.now().toString()
-    )
-    return false
-  }
-
-  if (lastOtpCheckString) {
-    return true
-  }
-
-  if (
-    createdAtString &&
-    createdAt &&
-    reminderCreatedAtDate &&
-    Date.now() > reminderCreatedAtDate
-  ) {
-    const resolve = await createOtpCheckModal()
-    if (resolve) {
-      await enableOtp(account)
-      await account.dataStore.setItem(
-        Constants.OTP_REMINDER_STORE_NAME,
-        Constants.OTP_REMINDER_KEY_NAME_LAST_OTP_CHECKED,
-        Date.now().toString()
-      )
-      return true
-    }
-    await account.dataStore.setItem(
-      Constants.OTP_REMINDER_STORE_NAME,
-      Constants.OTP_REMINDER_KEY_NAME_LAST_OTP_CHECKED,
-      Date.now().toString()
-    )
-    return false
-  }
-
-  return true
-}
-
-export const setMostRecentUsers = async (username: string) => {
-  const disklet = makeReactNativeDisklet()
-  const lastUsers = await disklet
-    .getText('lastusers.json')
-    .then(text => JSON.parse(text))
-    .catch(_ => [])
-  if (lastUsers && lastUsers.length > 0) {
-    const filteredLastUsers = lastUsers.filter(
-      (lastUser: string) => lastUser !== username
-    )
-    return disklet.setText(
-      'lastusers.json',
-      JSON.stringify([username, ...filteredLastUsers])
-    )
-  }
-  return disklet.setText('lastusers.json', JSON.stringify([username]))
 }
